Consolidate graph controls Cypress specs
